hf_space/requirements.txt
gradio>=5.47.0
ultralytics>=8.4.0
pillow>=10.0.0
